Move GasslessVotingToggle
