+++ src/pages/dao/settings/general/SafeGeneralSettingsPage.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { encodeFunctionData, zeroAddress } from 'viem';
 import { SettingsContentBox } from '../../../../components/SafeSettings/SettingsContentBox';
-import { GasslessVotingToggleDAOSettings } from '../../../../components/ui/cards/GasslessVotingToggleCard';
+import { GasslessVotingToggleDAOSettings } from '../../../../components/ui/GasslessVotingToggle';
 import { InputComponent } from '../../../../components/ui/forms/InputComponent';
 import { BarLoader } from '../../../../components/ui/loaders/BarLoader';
 import NestedPageHeader from '../../../../components/ui/page/Header/NestedPageHeader';
